@modify linux timer

+++ lib/libfaeris/src/sys/platform/linux/FsTimeLinux.cc
@@ -10,11 +10,11 @@ void Timer::reset()
 {
 	gettimeofday(&m_begin,NULL);
 }
-long Timer::now() const
+float Timer::now() const
 {
 	struct timeval cur;
 	gettimeofday(&cur,NULL);
-	return (cur.tv_sec-m_begin.tv_sec)*1000+(cur.tv_usec-m_begin.tv_usec)/1000;
+	return float(cur.tv_sec-m_begin.tv_sec)*1000.0f+float(cur.tv_usec-m_begin.tv_usec)/1000.0f;
 }
 
 NS_FS_END
